+++ llvm/lib/CodeGen/HardwareLoops.cpp
@@ -488,7 +488,7 @@ Value *HardwareLoop::InitLoopCount() {
 Value* HardwareLoop::InsertIterationSetup(Value *LoopCountInit) {
   IRBuilder<> Builder(BeginBB->getTerminator());
   if (BeginBB->getParent()->getAttributes().hasFnAttr(Attribute::StrictFP))
-    Builder.setIsFPConstrained(true);
+    Builder.setFPMode(true);
   Type *Ty = LoopCountInit->getType();
   bool UsePhi = UsePHICounter || Opts.ForcePhi;
   Intrinsic::ID ID = UseLoopGuard
@@ -518,7 +518,7 @@ void HardwareLoop::InsertLoopDec() {
   IRBuilder<> CondBuilder(ExitBranch);
   if (ExitBranch->getParent()->getParent()->getAttributes().hasFnAttr(
           Attribute::StrictFP))
-    CondBuilder.setIsFPConstrained(true);
+    CondBuilder.setFPMode(true);
 
   Value *Ops[] = { LoopDecrement };
   Value *NewCond = CondBuilder.CreateIntrinsic(Intrinsic::loop_decrement,
@@ -541,7 +541,7 @@ Instruction* HardwareLoop::InsertLoopRegDec(Value *EltsRem) {
   IRBuilder<> CondBuilder(ExitBranch);
   if (ExitBranch->getParent()->getParent()->getAttributes().hasFnAttr(
           Attribute::StrictFP))
-    CondBuilder.setIsFPConstrained(true);
+    CondBuilder.setFPMode(true);
 
   Value *Ops[] = { EltsRem, LoopDecrement };
   Value *Call = CondBuilder.CreateIntrinsic(Intrinsic::loop_decrement_reg,